Merged PR 15099: Using CommonTokenType.BLANK insted of SPACE_AS_DUMMY_TOKEN
Using CommonTokenType.BLANK instead of SPACE_AS_DUMMY_TOKEN

I did not like the solution with SPACE_AS_DUMMY_TOKEN very much before and going over it again I think I found a much better and clearer solution by forcing the BLANK as a mandatory token in any token enum. With this mandatory token it is more clear what happens. It then gets filtered out by the token factory instead of doing the filtering internally and not visible to the user/implementation.

I did not update the maven version as this is a non breaking change and we just started using the new version which is not even built with Drg-Commons yet.