Merged PR 33209: Implement AST parsing and evaluation.
Implement AST parsing and evaluation.

Add a new class DescentParser which will take a list of tokens and create
an abstract syntax tree from them. This tree can have any type of terminal
nodes, you just need to provide a transform Function<DrgLexerToken, your_type_here>
to the constructor. That function will be used to transform the token
into your type. Which also needs to be type defined on the parser itself.
See the test cases for sample usages.

When the DescentParser builds the tree, you can then evaluate the expression.
The result of the expression is an ExpressionMatch object which will
have the boolean result if the whole expression evaluation to true and
the matched objects of the specified type on the parser.

This can then be used to send messages via a bus or otherwise.