Merged PR 33448: Move evaluation out of DescentParser.
Move evaluation out of DescentParser.

Create new class DrgSyntaxTree. This class wraps the "root" of the tree
that was created by the parser. When the parser builds the expression
tree, it returns this new object. The evaluation logic now lives in the
DrgSyntaxTree. The the calling code can decide what
to do with the tree (e.g. evaluate or just cache the tree for later
evaluation or both!).

Add new setter to DescentParser, that allows a caller to set a new
tokenIterator so that the same parser can be used to parse more than one
list of tokens and be reused to create multiple ASTs.

Update tests and bump pom version number.