Merged PR 33619: Fix match collection with single list match.
Fix match collection with single list match.

When there is a single list as the formula, that means there is only
going to be a Terminal node in the AST. Currently we only add to the
matches collection in the Non-terminal nodes due to negation etc.

Update the evaluateExpression method to check if the tree evaluated to
true, is the root is a terminal and the matches are empty. If all this
is true, we can be confident that it was a single list AST and we can
add the root to the list of matches.

Bump version number