Merged PR 36318: Fix how the OR nodes was reporting matches.
Fix how the OR nodes was reporting matches.

If an OR node contains two terminal children and both were present in the
operands list, then both would evaluate to true. After evaluation, both
would be added to the matches, even though we only want to report the
left side as the match.

Add return paths after we add to the matches so that if we match the
right and are a terminal node, we just return.

Optimized the evaluation so that if we do match the right and are a
terminal, we return before evaluating the right. To save some processing
time in certain cases.

Replace RuntimeException with FarserException in DescentParser.java