Merged PR 36413: Add custom terminal node functionality.
Add custom terminal node functionality.

Change how the AST creates the terminal nodes such that a calling application
can now provide their own implementations of terminal nodes that
encapsulate more logic outside of a simple list#contains. The DescentParser
now takes a map of suppliers that it will use to create terminal nodes.
The parser will use the Atom token value to look suppliers up in the map.
If the map does not contain a supplier for a particular token value it
will default to a supplier that was provided in the constructor by
the calling application.

Remove the Terminal class and move those values into the ContainsNode.
Change where the matches are added to the accumulator to simplify the
operator Nodes.