Merged PR 46548: Merge updates to dev
Some dependency updates and some static lexer methods moved around. I talked to Melanie and she does not use those lexer methods so I did not increase the version since it is still a SNAPSHOT, but I can increase it if requested.

Exception with printing out token in `DescentParser` to have some hint on where it fails

Site version stuff updated and more lexer documentation and some initial AST documentation added.

`DrgFormulaLexerTest` contained tests for both the DRG and the Domain lexing - split up into two separate testing classes which each test everything.

`SimpleLexerTest` updated with some more complex token situations around token characters which also appear in another token.

Handling of optionals when looking for mandatory common token types in the custom token implementation. The existence of the mandatory tokens is checked at some point when it creates a token lookup map, but when accessing those tokens in the lexer it returns (returned) optionals - even though we know those are mandatory and will exist based on the knowledge of the check when it generates the lookup map. So, in the lexer there is no guarantee or indication that those tokens will exist, so the optionals checking has to happen there too.